og/requirements.txt
# Core dependencies
streamlit==1.29.0
pdfplumber==0.10.3
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
nltk==3.8.1
openai==1.6.1

# Additional dependencies
torch==2.1.0
transformers==4.35.2
google-generativeai==0.3.2
python-dotenv==1.0.0

# FastAPI backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
